RUN-415 Refactoring
