+++ Runar/Runar/Controllers/Misc/TabBarController/MainTabBarController.swift
@@ -74,25 +74,3 @@ class MainTabBarController: UITabBarController {
         return navigationVC
     }
 }
-
-class CustomIntensityVisualEffectView: UIVisualEffectView {
-    
-    private var animator: UIViewPropertyAnimator?
-    
-    init(effect: UIVisualEffect, intensity: CGFloat) {
-        super.init(effect: nil)
-        
-        animator = UIViewPropertyAnimator(duration: 1, curve: .linear) { [weak self] in
-            self?.effect = effect
-        }
-        animator?.fractionComplete = intensity
-        animator?.pausesOnCompletion = true
-    }
-
-    required init?(coder aDecoder: NSCoder) {
-        fatalError()
-    }
-}
-
-
-
